// frontend/src/models/Store.ts
export type StorePost = {
  Name: string;
  Email: string;
  PhoneNumber: string;
  Address: string;
  CoverPhoto: string;
  ProfilePhoto: string;
};

export type StoreGet = {
  id: number;
  name: string;
  email: string;
  phoneNumber: string;
  address: string;
  coverPhoto: string;
  profilePhoto: string;
};